Fixed PrinterHost demo Printer_GetDeviceID() routine not removing the Device ID string length from the start of the returned array (thanks to John Andrews).
Fixed error in new pipe stream function template system not setting the right device token for R/W operations (also thanks to John Andrews).

+++ Demos/Host/Incomplete/PrinterHost/Lib/PrinterCommands.h
@@ -62,7 +62,7 @@
 		
 	/* Function Prototypes: */
 		uint8_t Printer_SendData(Printer_Data_t* PrinterCommands);
-		uint8_t Printer_GetDeviceID(char* DeviceIDString, uint8_t BufferSize);
+		uint8_t Printer_GetDeviceID(char* DeviceIDString, uint16_t BufferSize);
 		uint8_t Printer_GetPortStatus(uint8_t* PortStatus);
 		uint8_t Printer_SoftReset(void);
